[shuffle] replace json rpc with api
