+++ shuffle/cli/src/account.rs
@@ -1,13 +1,13 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::shared::{send_transaction, Home, Network, NetworkHome, LOCALHOST_NAME};
+use crate::shared::{DevApiClient, Home, Network, NetworkHome, LOCALHOST_NAME};
 use anyhow::{anyhow, Result};
 use diem_crypto::PrivateKey;
 
 use diem_infallible::duration_since_epoch;
 use diem_sdk::{
-    client::{BlockingClient, FaucetClient},
+    client::FaucetClient,
     transaction_builder::{Currency, TransactionFactory},
     types::LocalAccount,
 };
@@ -49,15 +49,18 @@ pub async fn handle(home: &Home, root: Option<PathBuf>, network: Network) -> Res
         }
         None => {
             println!("Connecting to {}...", network.get_json_rpc_url());
-            let client = BlockingClient::new(network.get_json_rpc_url());
+            let client = DevApiClient::new(reqwest::Client::new(), network.get_dev_api_url())?;
             let factory = TransactionFactory::new(ChainId::test());
 
             if let Some(input_root_key) = root {
                 network_home.copy_key_to_latest(input_root_key.as_path())?
             }
-            let mut treasury_account = get_treasury_account(&client, home.get_root_key_path())?;
-            create_account_via_dev_api(&mut treasury_account, &new_account, &factory, &client)?;
+            let mut treasury_account =
+                get_treasury_account(&client, home.get_root_key_path()).await?;
+            create_account_via_dev_api(&mut treasury_account, &new_account, &factory, &client)
+                .await?;
             create_account_via_dev_api(&mut treasury_account, &test_account, &factory, &client)
+                .await
         }
     }
 }
@@ -134,13 +137,14 @@ fn generate_test_account(network_home: &NetworkHome) -> Result<LocalAccount> {
     ))
 }
 
-pub fn get_treasury_account(client: &BlockingClient, root_key_path: &Path) -> Result<LocalAccount> {
+pub async fn get_treasury_account(
+    client: &DevApiClient,
+    root_key_path: &Path,
+) -> Result<LocalAccount> {
     let treasury_account_key = load_key(root_key_path);
     let treasury_seq_num = client
-        .get_account(account_config::treasury_compliance_account_address())?
-        .into_inner()
-        .unwrap()
-        .sequence_number;
+        .get_account_sequence_number(account_config::treasury_compliance_account_address())
+        .await?;
     Ok(LocalAccount::new(
         account_config::treasury_compliance_account_address(),
         treasury_account_key,
@@ -148,35 +152,34 @@ pub fn get_treasury_account(client: &BlockingClient, root_key_path: &Path) -> Re
     ))
 }
 
-pub fn create_account_via_dev_api(
+pub async fn create_account_via_dev_api(
     treasury_account: &mut LocalAccount,
     new_account: &LocalAccount,
     factory: &TransactionFactory,
-    client: &BlockingClient,
+    client: &DevApiClient,
 ) -> Result<()> {
-    println!("Creating a new account onchain...");
-    if client
-        .get_account(new_account.address())
-        .unwrap()
-        .into_inner()
-        .is_some()
-    {
-        println!("Account already exists: {}", new_account.address());
-    } else {
-        let create_new_account_txn = treasury_account.sign_with_transaction_builder(
-            factory.payload(encode_create_parent_vasp_account_script_function(
-                Currency::XUS.type_tag(),
-                0,
-                new_account.address(),
-                new_account.authentication_key().prefix().to_vec(),
-                vec![],
-                false,
-            )),
-        );
-        send_transaction(client, create_new_account_txn)?;
-        println!("Successfully created account {}", new_account.address());
-    }
-    println!("Public key: {}", new_account.public_key());
+    match client.get_account_resources(new_account.address()).await {
+        Ok(_) => println!("Account already exists: {}", new_account.address()),
+        Err(_) => {
+            println!("Creating a new account onchain...");
+            let create_new_account_txn = treasury_account.sign_with_transaction_builder(
+                factory.payload(encode_create_parent_vasp_account_script_function(
+                    Currency::XUS.type_tag(),
+                    0,
+                    new_account.address(),
+                    new_account.authentication_key().prefix().to_vec(),
+                    vec![],
+                    false,
+                )),
+            );
+            let bytes = bcs::to_bytes(&create_new_account_txn)?;
+            let json = client.post_transactions(bytes).await?;
+            let hash = DevApiClient::get_hash_from_post_txn(json)?;
+            client.check_txn_executed_from_hash(hash.as_str()).await?;
+            println!("Successfully created account {}", new_account.address());
+            println!("Public key: {}", new_account.public_key());
+        }
+    };
     Ok(())
 }
