+++ shuffle/integration-tests/src/lib.rs
@@ -6,8 +6,11 @@ mod helper;
 use crate::helper::ShuffleTestHelper;
 use forge::{AdminContext, AdminTest, Result, Test};
 use move_cli::package::cli::UnitTestResult;
+use shuffle::shared::DevApiClient;
 use smoke_test::scripts_and_modules::enable_open_publishing;
+use std::str::FromStr;
 use tokio::runtime::Runtime;
+use url::Url;
 
 pub struct SamplePackageEndToEnd;
 
@@ -61,6 +64,10 @@ impl AdminTest for TypescriptSdkIntegration {
 
 fn bootstrap_shuffle(ctx: &mut AdminContext<'_>) -> Result<ShuffleTestHelper> {
     let client = ctx.client();
+    let dev_api_client = DevApiClient::new(
+        reqwest::Client::new(),
+        Url::from_str(ctx.chain_info().rest_api())?,
+    )?;
     let factory = ctx.chain_info().transaction_factory();
     enable_open_publishing(&client, &factory, ctx.chain_info().root_account())?;
 
@@ -70,11 +77,10 @@ fn bootstrap_shuffle(ctx: &mut AdminContext<'_>) -> Result<ShuffleTestHelper> {
     // let mut account = ctx.random_account(); // TODO: Support arbitrary addresses
     let mut account = ShuffleTestHelper::hardcoded_0x2416_account(&client)?;
     let tc = ctx.chain_info().treasury_compliance_account();
-    helper.create_account(tc, &account, factory, client)?;
-
     let rt = Runtime::new().unwrap();
     let handle = rt.handle().clone();
-    handle.block_on(helper.deploy_project(&mut account, ctx.chain_info().rest_api()))?;
 
+    handle.block_on(helper.create_account(tc, &account, factory, &dev_api_client))?;
+    handle.block_on(helper.deploy_project(&mut account, ctx.chain_info().rest_api()))?;
     Ok(helper)
 }
